simple-data-platform/pyproject.toml
[tool.poetry]
name = "simple-data-platform"
version = "0.1.0"
description = ""
authors = ["Burak Basaran <basaranoguzburak@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.28.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.frontend.dependencies]
bokeh = "^3.6.2"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"


[tool.poetry.group.prefectworker.dependencies]
pandas = "^2.2.3"


[tool.poetry.group.prefect.dependencies]
prefect = "2.19.5"
# later anyio versions break prefect
anyio = "3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
